+++ src/routes/details/ocpBreakdown/providerDetails/clusterInfo/clusterInfo.styles.ts
