+++ src/routes/details/ocpBreakdown/providerDetails/clusterInfo/clusterInfo.tsx
