+++ src/routes/details/ocpBreakdown/providerDetails/clusterInfo/clusterInfoContent.scss
