+++ src/routes/details/ocpBreakdown/providerDetails/clusterInfo/index.ts
