+++ src/routes/details/ocpBreakdown/providerDetails/components/sourceLink.styles.ts
@@ -1,0 +1,8 @@
+import global_spacer_md from '@patternfly/react-tokens/dist/js/global_spacer_md';
+import type React from 'react';
+
+export const styles = {
+  spacingRight: {
+    marginRight: global_spacer_md.value,
+  },
+} as { [className: string]: React.CSSProperties };
