+++ src/routes/details/ocpBreakdown/providerDetails/dataDetails/components/costData.tsx
@@ -3,10 +3,10 @@ import type { Provider } from 'api/providers';
 import messages from 'locales/messages';
 import React from 'react';
 import { useIntl } from 'react-intl';
-import { styles } from 'routes/details/ocpBreakdown/dataDetails/dataDetails.styles';
-import { formatDate } from 'routes/details/ocpBreakdown/dataDetails/utils/format';
-import { getProgressStepIcon } from 'routes/details/ocpBreakdown/dataDetails/utils/icon';
-import { getProgressStepVariant } from 'routes/details/ocpBreakdown/dataDetails/utils/variant';
+import { styles } from 'routes/details/ocpBreakdown/providerDetails/dataDetails/dataDetails.styles';
+import { formatDate } from 'routes/details/ocpBreakdown/providerDetails/dataDetails/utils/format';
+import { getProgressStepIcon } from 'routes/details/ocpBreakdown/providerDetails/dataDetails/utils/icon';
+import { getProgressStepVariant } from 'routes/details/ocpBreakdown/providerDetails/dataDetails/utils/variant';
 
 interface CostDataOwnProps {
   provider: Provider;
