+++ src/routes/details/ocpBreakdown/providerDetails/dataDetails/dataDetailsContent.tsx
@@ -8,9 +8,9 @@ import { useIntl } from 'react-intl';
 import { useSelector } from 'react-redux';
 import { NotAvailable } from 'routes/components/page/notAvailable';
 import { LoadingState } from 'routes/components/state/loadingState';
-import { CloudIData } from 'routes/details/ocpBreakdown/dataDetails/components/cloudIData';
-import { ClusterData } from 'routes/details/ocpBreakdown/dataDetails/components/clusterData';
-import { CostData } from 'routes/details/ocpBreakdown/dataDetails/components/costData';
+import { CloudIData } from 'routes/details/ocpBreakdown/providerDetails/dataDetails/components/cloudIData';
+import { ClusterData } from 'routes/details/ocpBreakdown/providerDetails/dataDetails/components/clusterData';
+import { CostData } from 'routes/details/ocpBreakdown/providerDetails/dataDetails/components/costData';
 import { filterProviders } from 'routes/utils/providers';
 import type { RootState } from 'store';
 import { FetchStatus } from 'store/common';
