+++ src/routes/details/ocpBreakdown/providerDetails/dataDetails/index.ts
