+++ src/routes/details/ocpBreakdown/providerDetails/dataDetails/utils/format.ts
