+++ src/routes/details/ocpBreakdown/providerDetails/dataDetails/utils/icon.tsx
