+++ src/routes/details/ocpBreakdown/providerDetails/dataDetails/utils/variant.ts
