+++ src/routes/details/ocpBreakdown/providerDetails/utils/normailize.ts
@@ -1,0 +1,3 @@
+export const normalize = (value: string) => {
+  return value ? value.toLowerCase().replace('-', '_') : undefined;
+};
